OS-7295. Move WebkitTextSecurity to masked class (#546)
The style for the "credentials" field was modified in OS-6745.

Specifically, the -webkit-text-security property was relocated from input element classes to the Input component styles. However, this adjustment unexpectedly disrupted the spacing calculation for labels within the Input component.

The rationale behind transferring the rule from classes to styles is unclear to me. Upon inspection in both Chrome and Firefox, I couldn't discern any discernible differences between the two variations – the field text is covered with dots.

Consequently, I've reverted the rule back to classes to ensure its application to the input element.