OS-7716. Add support for getting ML task tags.
This commit adds support for retrieving ML task tags in the API. The `getMlTaskTags` function is implemented in the `api/restapi/actionCreators.ts` file to make a GET request to the appropriate endpoint. The response is then handled and stored in the state using the `SET_ML_TASK_TAGS` action type. This functionality will allow users to retrieve tags associated with ML tasks.